Replaced AST comparison of Java files with single line comparison
Comparing complete file ASTs was slow, so changed it to single line comparison based on Spoon's pretty-printer or Gumtree Spoon AST Diff's snippet compare. To this end, built a simple class template and used JPype to interact with Spoon and Gumtree Spoon AST Diff for single line comparison.

Because of using JPype objects, `joblib` can't `pickle` them for the parallel execution. So, replaced it with `pathos` that uses `dill` instead of `pickle` and is ok with JPype objects.

Finally, while I was at it, used `tqdm` for the looks!